更改关系表（一分为三），相关函数更改至还剩共同关注/粉丝

+++ models/group.js
@@ -1,7 +1,7 @@
 module.exports = (sequelize, DataTypes) => sequelize.define('Group', {
   id: {
     allowNull: false,
-    type: DataTypes.INTEGER(100),
+    type: DataTypes.INTEGER(255),
     autoIncrement: true,
     primaryKey: true,
     comment: '分组id'
